began building out home page components using rebass
