+++ gatsby-config.js
@@ -6,6 +6,7 @@ module.exports = {
   },
 
   plugins: [
+    `gatsby-plugin-theme-ui`,
     `gatsby-plugin-react-helmet`,
     {
       resolve: `gatsby-plugin-create-client-paths`,
