+++ package.json
@@ -7,6 +7,7 @@
     "@reach/dialog": "^0.2.9",
     "@reach/tabs": "^0.1.6",
     "@reach/visually-hidden": "^0.1.4",
+    "@rebass/preset": "^4.0.5",
     "axios": "^0.19.2",
     "babel-eslint": "^10.1.0",
     "dotenv": "^8.2.0",
@@ -22,6 +23,7 @@
     "gatsby-plugin-offline": "^2.2.10",
     "gatsby-plugin-react-helmet": "^3.1.6",
     "gatsby-plugin-sharp": "^2.2.20",
+    "gatsby-plugin-theme-ui": "^0.3.0",
     "gatsby-source-filesystem": "^2.1.21",
     "gatsby-transformer-json": "^2.4.5",
     "gatsby-transformer-sharp": "^2.2.13",
@@ -33,9 +35,12 @@
     "prop-types": "^15.7.2",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
+    "react-hamburger-menu": "^1.2.1",
     "react-helmet": "^5.2.1",
     "react-netlify-identity-widget": "^0.2.5",
-    "rss-parser": "^3.8.0"
+    "rebass": "^4.0.7",
+    "rss-parser": "^3.8.0",
+    "theme-ui": "^0.3.1"
   },
   "keywords": [],
   "license": "MIT",
